# panel/openapi.yaml
openapi: "3.0.3"
info:
  title: SBoard Panel API
  version: "1.0.0"
  description: SBoard panel management API
servers:
  - url: /api
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    BootstrapStatus:
      type: object
      required:
        - needs_setup
      properties:
        needs_setup:
          type: boolean
    BootstrapResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
...
    BootstrapRequest:
      type: object
      required:
        - username
        - password
        - confirm_password
      properties:
        username:
          type: string
        password:
          type: string
        confirm_password:
          type: string
        setup_token:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /admin/bootstrap:
    get:
      operationId: getBootstrapStatus
      summary: Check if admin setup is needed
      security: []
      responses:
        "200":
          description: Bootstrap status
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/BootstrapStatus"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: bootstrap
      summary: Create initial admin account
      security: []
      parameters:
        - name: X-Setup-Token
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapRequest"
      responses:
        "200":
          description: Admin created
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/BootstrapResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Auth ────────────────────────────────────────────────
...
  /sub/{user_uuid}:
    get:
      operationId: getSubscription
      summary: Get user subscription
      security: []
      parameters:
        - name: user_uuid
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum:
              - singbox
              - v2ray
      responses:
        "200":
          description: Subscription content
          content:
            application/json:
              schema:
                type: object
                description: sing-box configuration
            text/plain:
              schema:
                type: string
                description: v2ray subscription (base64 encoded links)
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Users ───────────────────────────────────────────────
...